requirements.txt
asyncpraw==7.7.0
asyncprawcore==2.3.0
beautifulsoup4==4.12.2
discord.py==2.3.1
gpt4all==1.0.8
librosa==0.9.2
openai==0.27.8
Pillow==10.0.0
Pillow==10.2.0
psutil==5.9.5
Requests==2.31.0
youtube_dl==2021.12.17
